cpuclass: Add CPU class support
This patch enables CPU class support for ARM and x86 platforms.
It includes basic hypervisor and toolstack support, and targets
the ARM big.LITTLE architecture.

There is a placeholder for the x86 case, although it should be
fixed/removed.

Essentially it is an initial implementation of the design document:
"Heterogeneous Multi Processing Support in Xen".

Signed-off-by: Thanasis Katsios <thkatsios@cslab.ece.ntua.gr>
Signed-off-by: Anastassios Nanos <ananos@cslab.ece.ntua.gr>